Add BuildCPanel hook with extended DForm functionality.
DForm just wasn't cutting it in terms of what I needed out of the interface of the tool. So I created extended functionality to DForm that takes the DForm as an input and adds panels to it. Rather than trying to override the default functionality and potentially meddle with other tools, this was an easier and closer happy medium to what I wanted in terms of a layout. Unfinished though.